:wrench: Fixed all msvc level 3 warnings.

+++ src/Plugin.cpp
@@ -128,11 +128,12 @@ void Plugin::init() {
 
   mGuiManager->getGui()->registerCallback("atmosphere.setQuality",
       "Higher values create a more realistic atmosphere.",
-      std::function([this](double value) { mProperties->mQuality = value; }));
+      std::function([this](double value) { mProperties->mQuality = static_cast<int>(value); }));
 
   mGuiManager->getGui()->registerCallback("atmosphere.setWaterLevel",
       "Sets the height of the water surface relative to the planet's radius.",
-      std::function([this](double value) { mProperties->mWaterLevel = value; }));
+      std::function(
+          [this](double value) { mProperties->mWaterLevel = static_cast<float>(value); }));
 
   mEnableShadowsConnection = mGraphicsEngine->pEnableShadows.connectAndTouch([this](bool val) {
     for (auto const& atmosphere : mAtmospheres) {
@@ -222,7 +223,7 @@ void Plugin::update() {
 
     auto sunDirection = mSolarSystem->getSunDirection(atmosphere->getWorldTransform()[3]);
 
-    atmosphere->getRenderer().setSun(sunDirection, sunIlluminance);
+    atmosphere->getRenderer().setSun(sunDirection, static_cast<float>(sunIlluminance));
   }
 }
